requirements.txt
requests-oauthlib==1.3.0
requests==2.25.0
wtforms==2.3.3
python-dateutil==2.8.1
pytz==2020.4
Flask==1.1.2
gunicorn==20.0.4
psycopg2-binary==2.8.6
flask-talisman==0.7.0
flask_login==0.5.0
flask_migrate==2.5.3
flask_wtf==0.14.3
sentry-sdk[flask]==0.19.4
mock==4.0.2
py-trello==0.17.1
jsmin==2.2.2
tmdbsimple==2.6.6
email_validator==1.1.2
pytest==6.1.2
mypy==0.790
Jinja2==2.11.2
python-pushover==0.4
python-youtube==0.6.4
isodate==0.6.0
marshmallow==3.9.1
sendgrid==6.4.7
humanize==3.1.0
tld==0.12.2
pifx==1.0.1
suntime==1.2.5
b2sdk==1.2.0
spacy==2.3.2
wikipedia==1.4.0
spotipy==2.16.1
black==20.8b1

https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.0/en_core_web_sm-2.3.0.tar.gz
git+git://github.com/z1lc/genanki.git#egg=genanki
